% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecPlot.R
\name{SpecPlotter}
\alias{SpecPlotter}
\title{Field Plotting Function}
\usage{
SpecPlotter(
  FolderOfInterest,
  Filename,
  FileExtension = "ssm",
  Title = Filename,
  Ylab = "Count"
)
}
\arguments{
\item{FolderOfInterest}{A character string providing the path to the directory housing the file of interest.}

\item{Filename}{A character string providing the file name for the file of interest WITHOUT EXTENSION.}

\item{FileExtension}{A character string providing the file extension of the file of interest WITHOUT THE LEADING DOT. This defaults to "ssm" which corresponds with scope mode.}

\item{Title}{A character string providing the desired plot title. Defaults to the file name.}

\item{Ylab}{A character string providing the desired y-axis label. Defaults to "Count" which is appropriate for data collected in scope mode.}
}
\description{
This function creates a dark-themed plot for quickly checking exported spectral data while maintaining the integrity of your scoptopic vision.
}
\examples{
#Plotting a .ssm file
SpecPlotter("~/Documents", "Test_File")

#Plotting a .irr file
SpecPlotter("~/Documents", "Test_File2", "irr")
}
\keyword{plotting}
